Refactor tests to pass CI
