+++ tests/integration/conftest.py
@@ -6,8 +6,11 @@ import logging
 from pathlib import Path
 
 import jubilant
+import lightkube
 import pytest
 import yaml
+from lightkube.models.meta_v1 import ObjectMeta
+from lightkube.resources.core_v1 import Namespace
 from typing_extensions import Literal
 
 logger = logging.getLogger(__name__)
@@ -15,6 +18,11 @@ logging.getLogger("jubilant.wait").setLevel(logging.WARNING)
 
 METADATA = yaml.safe_load(Path("./metadata.yaml").read_text())
 APP_NAME = METADATA["name"]
+
+
+KUBEFLOW_USER_PROFILE_A = "kubeflow-profile-a"
+KUBEFLOW_USER_PROFILE_B = "kubeflow-profile-b"
+RESOURCE_DISPATCHER = "resource-dispatcher"
 
 
 def pytest_addoption(parser):
@@ -93,3 +101,53 @@ def juju_vm(request: pytest.FixtureRequest, vm_controller: str):
         if request.session.testsfailed:
             log = juju.debug_log(limit=30)
             print(log, end="")
+
+
+@pytest.fixture(scope="session")
+def lightkube_client() -> lightkube.Client:
+    client = lightkube.Client(field_manager=RESOURCE_DISPATCHER)
+    return client
+
+
+@pytest.fixture(scope="module")
+def kubeflow_user_profile_a(lightkube_client: lightkube.Client):
+    """Return a new namespace with the label user.kubeflow.org/enabled=true."""
+    namespace_name = KUBEFLOW_USER_PROFILE_A
+    namespace = Namespace(
+        metadata=ObjectMeta(
+            name=namespace_name,
+            labels={
+                "user.kubeflow.org/enabled": "true",
+                "app.kubernetes.io/part-of": "kubeflow-profile",
+            },
+        )
+    )
+    logger.info(
+        f"Creating namespace {namespace_name} with label user.kubeflow.org/enabled=true ..."
+    )
+    lightkube_client.create(namespace)
+    assert namespace.metadata
+    yield namespace.metadata.name
+    lightkube_client.delete(Namespace, name=namespace_name)
+
+
+@pytest.fixture(scope="module")
+def kubeflow_user_profile_b(lightkube_client: lightkube.Client):
+    """Return a new namespace with the label user.kubeflow.org/enabled=true."""
+    namespace_name = KUBEFLOW_USER_PROFILE_B
+    namespace = Namespace(
+        metadata=ObjectMeta(
+            name=namespace_name,
+            labels={
+                "user.kubeflow.org/enabled": "true",
+                "app.kubernetes.io/part-of": "kubeflow-profile",
+            },
+        )
+    )
+    logger.info(
+        f"Creating namespace {namespace_name} with label user.kubeflow.org/enabled=true ..."
+    )
+    lightkube_client.create(namespace)
+    assert namespace.metadata
+    yield namespace.metadata.name
+    lightkube_client.delete(Namespace, name=namespace_name)
